refactor: extract editor's timeline into its own component
